aligning the definition of grouplets
We already had a very clear distinction between guilds and working groups, that should be reflected here so new entrants can understand. Grouplet is the "super-set" with WGs being "loose" right now and Guilds being "structured". We might create other types of grouplets in the future, so leaving that open-ended.

Also small edits for active voice, plain language, and length